getlocation frontend works, getting city, country from google api also, now working on getlocation and getting the delayed data to the back-end.. need to find a solution still
