+++ package.json
@@ -10,6 +10,7 @@
     "express-session": "^1.14.2",
     "pg": "^6.1.0",
     "pug": "^2.0.0-beta6",
+    "request": "^2.79.0",
     "sequelize": "^3.27.0"
   },
   "devDependencies": {},
