+++ views/includes/footer.pug
@@ -1,0 +1,1 @@
+script(type='text/javascript' src='/js/getlocation.js')
